preliminary work on 3.0-series
