+++ lib/log.js
@@ -1,0 +1,10 @@
+(function() {
+	'use strict';
+
+	module.exports = function() {
+		if (!!process.env.PIPEDRIVE_DEBUG) {
+			console.log.apply(this, arguments);
+		}
+	};
+
+})();
